refactor(timetree): extract domain helpers from commands/ to timetree module
- Move TimeMarginalMode enum to timetree/params.rs (domain concept, not CLI-specific)
- Move compute_effective_time_marginal and build_covariation_clock_params to timetree/params.rs with individual-param signatures instead of TreetimeTimetreeArgs dependency
- Move run_refinement_iteration to timetree/refinement.rs with RefinementParams struct
- commands/timetree/ becomes thin adapter constructing params from args
