+++ packages/treetime/src/timetree/mod.rs
@@ -2,4 +2,6 @@ pub mod confidence;
 pub mod convergence;
 pub mod inference;
 pub mod optimization;
+pub mod params;
+pub mod refinement;
 pub mod utils;
